refactor (BattleState): moved back BattleState in its own file
